+++ Assets/Tests/InputSystem/Plugins/UITests.cs
@@ -3277,14 +3277,18 @@ internal class UITests : CoreTestsFixture
     #if UNITY_2021_2_OR_NEWER
     [UnityTest]
     [Category("UI")]
+    [TestCase(UIPointerBehavior.AllPointersAsIs, ExpectedResult = 1)]
+    [TestCase(UIPointerBehavior.SingleMouseOrPenButMultiTouchAndTrack, ExpectedResult = 1)]
+    [TestCase(UIPointerBehavior.SingleUnifiedPointer, ExpectedResult = 1)]
     #if UNITY_ANDROID || UNITY_IOS || UNITY_TVOS
     [Ignore("Currently fails on the farm but succeeds locally on Note 10+; needs looking into.")]
     #endif
     [PrebuildSetup(typeof(UI_CanOperateUIToolkitInterface_UsingInputSystemUIInputModule_Setup))]
-    public IEnumerator UI_CanOperateUIToolkitInterface_UsingInputSystemUIInputModule()
+    public IEnumerator UI_CanOperateUIToolkitInterface_UsingInputSystemUIInputModule(UIPointerBehavior pointerBehavior)
     {
         var mouse = InputSystem.AddDevice<Mouse>();
         var gamepad = InputSystem.AddDevice<Gamepad>();
+        InputSystem.AddDevice<Touchscreen>();
 
         var scene = SceneManager.LoadScene("UITKTestScene", new LoadSceneParameters(LoadSceneMode.Additive));
         yield return null;
@@ -3300,6 +3304,8 @@ internal class UITests : CoreTestsFixture
             var uiButton = uiRoot.Query<UnityEngine.UIElements.Button>("Button").First();
             var scrollView = uiRoot.Query<ScrollView>("ScrollView").First();
 
+            uiModule.pointerBehavior = pointerBehavior;
+
             var clickReceived = false;
             uiButton.clicked += () => clickReceived = true;
             // NOTE: We do *NOT* do the following as the gamepad submit action will *not* trigger a ClickEvent.
@@ -3308,6 +3314,7 @@ internal class UITests : CoreTestsFixture
             yield return null;
 
             var buttonCenter = new Vector2(uiButton.worldBound.center.x, Screen.height - uiButton.worldBound.center.y);
+            var buttonOutside = new Vector2(uiButton.worldBound.center.x, Screen.height - (uiButton.worldBound.max.y + 10));
             var scrollViewCenter = new Vector2(scrollView.worldBound.center.x, Screen.height - scrollView.worldBound.center.y);
 
             Set(mouse.position, buttonCenter, queueEventOnly: true);
@@ -3353,6 +3360,30 @@ internal class UITests : CoreTestsFixture
             Assert.That(clickReceived, Is.True);
 
             ////TODO: tracked device support (not yet supported by UITK)
+
+            static bool IsActive(VisualElement ve)
+            {
+                return ve.Query<VisualElement>().Active().ToList().Contains(ve);
+            }
+
+            // Case 1369081: Make sure button doesn't get "stuck" in an active state when multiple fingers are used.
+            BeginTouch(1, buttonCenter, queueEventOnly: true);
+            yield return null;
+            Assert.That(IsActive(uiButton), Is.True);
+
+            BeginTouch(2, buttonOutside, queueEventOnly: true);
+            yield return null;
+            EndTouch(2, buttonOutside, queueEventOnly: true);
+            yield return null;
+
+            if (pointerBehavior == UIPointerBehavior.SingleUnifiedPointer)
+                Assert.That(IsActive(uiButton), Is.False);
+            else
+                Assert.That(IsActive(uiButton), Is.True);
+
+            EndTouch(1, buttonCenter, queueEventOnly: true);
+            yield return null;
+            Assert.That(IsActive(uiButton), Is.False);
         }
         finally
         {